# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: PrepVI R1 API
  version: 1.0.0
  description: Same-origin API contract for PrepVI US-01–30. Sensitive fields are intentionally absent from error schemas.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /job-descriptions:
    get:
      operationId: listJobDescriptions
      responses: { "200": { $ref: "#/components/responses/Page" } }
    post:
      operationId: createJobDescription
      requestBody:
        required: true
        content:
          application/json: { schema: { type: object, required: [text], properties: { text: { type: string, maxLength: 50000 } } } }
          multipart/form-data: { schema: { type: object, required: [file], properties: { file: { type: string, format: binary } } } }
      responses: { "201": { $ref: "#/components/responses/JobDescription" }, "413": { $ref: "#/components/responses/Error" }, "415": { $ref: "#/components/responses/Error" } }
  /job-descriptions/{jobDescriptionId}:
    parameters: [ { $ref: "#/components/parameters/JobDescriptionId" } ]
    get:
      operationId: getJobDescription
      responses: { "200": { $ref: "#/components/responses/JobDescription" }, "404": { $ref: "#/components/responses/Error" } }
  /job-descriptions/{jobDescriptionId}/extract:
    parameters: [ { $ref: "#/components/parameters/JobDescriptionId" }, { $ref: "#/components/parameters/IdempotencyKey" } ]
    post:
      operationId: startExtraction
      responses: { "202": { $ref: "#/components/responses/JobDescription" }, "409": { $ref: "#/components/responses/Error" } }
  /job-descriptions/{jobDescriptionId}/extraction-retries:
    parameters: [ { $ref: "#/components/parameters/JobDescriptionId" }, { $ref: "#/components/parameters/IdempotencyKey" } ]
    post:
      operationId: retryExtraction
      responses: { "202": { $ref: "#/components/responses/JobDescription" } }
  /job-descriptions/{jobDescriptionId}/text:
    parameters: [ { $ref: "#/components/parameters/JobDescriptionId" } ]
    patch:
      operationId: saveCorrectedText
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/JobDescription" }, "409": { $ref: "#/components/responses/Error" } }
  /job-descriptions/{jobDescriptionId}/confirm:
    parameters: [ { $ref: "#/components/parameters/JobDescriptionId" } ]
    post:
      operationId: confirmCorrectedText
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/JobDescription" }, "409": { $ref: "#/components/responses/Error" } }
  /job-descriptions/{jobDescriptionId}/analyze:
    parameters: [ { $ref: "#/components/parameters/JobDescriptionId" }, { $ref: "#/components/parameters/IdempotencyKey" } ]
    post:
      operationId: analyzeJobDescription
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/Success" } }
  /job-descriptions/{jobDescriptionId}/analysis:
    parameters: [ { $ref: "#/components/parameters/JobDescriptionId" } ]
    get:
      operationId: getJobDescriptionAnalysis
      parameters: [ { name: analysisVersion, in: query, schema: { type: integer, minimum: 1 } } ]
      responses: { "200": { $ref: "#/components/responses/Success" } }
  /job-descriptions/{jobDescriptionId}/requirement-normalizations:
    parameters: [ { $ref: "#/components/parameters/JobDescriptionId" } ]
    put:
      operationId: saveRequirementNormalizations
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/Success" } }
  /job-descriptions/{jobDescriptionId}/matches:
    parameters: [ { $ref: "#/components/parameters/JobDescriptionId" } ]
    get:
      operationId: getMatches
      parameters: [ { name: analysisVersion, in: query, schema: { type: integer, minimum: 1 } } ]
      responses: { "200": { $ref: "#/components/responses/Success" } }
    post:
      operationId: createMatches
      parameters: [ { $ref: "#/components/parameters/IdempotencyKey" } ]
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/Success" } }
  /preparation-plans:
    get:
      operationId: listPreparationPlans
      responses: { "200": { $ref: "#/components/responses/Page" } }
    post:
      operationId: createPreparationPlan
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "201": { $ref: "#/components/responses/Success" } }
  /preparation-plans/{planId}:
    parameters: [ { $ref: "#/components/parameters/PlanId" } ]
    get:
      operationId: getPreparationPlan
      responses: { "200": { $ref: "#/components/responses/Success" }, "404": { $ref: "#/components/responses/Error" } }
  /preparation-plans/{planId}/items/{itemId}:
    parameters:
      - { $ref: "#/components/parameters/PlanId" }
      - { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      operationId: updatePreparationPlanItem
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/Success" }, "409": { $ref: "#/components/responses/Error" } }
  /preparation-plans/{planId}/mentor-candidates:
    parameters:
      - { $ref: "#/components/parameters/PlanId" }
      - { name: availableFrom, in: query, schema: { type: string, format: date-time } }
      - { name: availableTo, in: query, schema: { type: string, format: date-time } }
      - { name: page, in: query, schema: { type: integer, minimum: 1 } }
      - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
    get:
      operationId: listPreparationPlanMentorCandidates
      responses: { "200": { $ref: "#/components/responses/Page" } }
...
  /notifications/{notificationId}/read:
    parameters: [ { name: notificationId, in: path, required: true, schema: { type: string, format: uuid } } ]
    post:
      operationId: readNotification
      responses: { "200": { $ref: "#/components/responses/Success" } }
  /admin/operation-cases:
    get:
      operationId: listOperationCases
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [OPEN, IN_PROGRESS, RESOLVED, DISMISSED] } }
        - { name: type, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses: { "200": { $ref: "#/components/responses/Page" } }
  /admin/operation-cases/{caseId}:
    parameters: [ { $ref: "#/components/parameters/CaseId" } ]
    get:
      operationId: getOperationCase
      responses: { "200": { $ref: "#/components/responses/Success" } }
  /admin/operation-cases/{caseId}/impact:
    parameters: [ { $ref: "#/components/parameters/CaseId" } ]
    get:
      operationId: previewOperationImpact
      responses: { "200": { $ref: "#/components/responses/Success" } }
  /admin/operation-cases/{caseId}/actions:
    parameters: [ { $ref: "#/components/parameters/CaseId" }, { $ref: "#/components/parameters/IdempotencyKey" } ]
    post:
      operationId: actOnOperationCase
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/Success" }, "409": { $ref: "#/components/responses/Error" }, "422": { $ref: "#/components/responses/Error" } }
  /admin/audit:
    get:
      operationId: listAuditLog
      parameters:
        - { name: targetType, in: query, schema: { type: string } }
        - { name: targetId, in: query, schema: { type: string, format: uuid } }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses: { "200": { $ref: "#/components/responses/Page" } }
  /admin/reports:
    get:
      operationId: listReports
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [OPEN, IN_REVIEW, RESOLVED, DISMISSED] } }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses: { "200": { $ref: "#/components/responses/Page" } }
...
  /admin/question-imports/preview:
    post:
      operationId: previewQuestionImport
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties: { file: { type: string, format: binary } }
      responses: { "201": { $ref: "#/components/responses/Success" }, "422": { $ref: "#/components/responses/Error" } }
  /admin/question-imports/{importId}:
    parameters: [ { name: importId, in: path, required: true, schema: { type: string, format: uuid } } ]
    get:
      operationId: getQuestionImport
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [VALID, INVALID, IMPORTED, SKIPPED] } }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses: { "200": { $ref: "#/components/responses/Success" } }
  /admin/question-imports/{importId}/commit:
    parameters:
      - { name: importId, in: path, required: true, schema: { type: string, format: uuid } }
      - { $ref: "#/components/parameters/IdempotencyKey" }
    post:
      operationId: commitQuestionImport
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/Success" }, "409": { $ref: "#/components/responses/Error" } }
  /admin/question-imports/{importId}/errors.csv:
    parameters: [ { name: importId, in: path, required: true, schema: { type: string, format: uuid } } ]
    get:
      operationId: downloadQuestionImportErrors
      responses:
        "200":
          description: CSV containing invalid rows and recovery instructions
          content: { text/csv: { schema: { type: string } } }
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: __Host-prepvi_session }
  parameters:
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 200 } }
...
    JobDescriptionId: { name: jobDescriptionId, in: path, required: true, schema: { type: string, format: uuid } }
    PlanId: { name: planId, in: path, required: true, schema: { type: string, format: uuid } }
...
    CaseId: { name: caseId, in: path, required: true, schema: { type: string, format: uuid } }
  requestBodies:
    JsonInput:
      required: true
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
  schemas:
    Recovery:
      type: object
      required: [kind, retryable, retryAfterSeconds]
      properties:
        kind: { type: string, enum: [RETRY_SAFE, EDIT_INPUT, REUPLOAD, PASTE_TEXT, SELECT_ANOTHER_SLOT, WAIT, CONTACT_SUPPORT, NONE] }
        retryable: { type: boolean }
        retryAfterSeconds: { type: [integer, "null"] }
    Error:
      type: object
      additionalProperties: false
      required: [code, message, correlationId, fieldErrors, recovery]
      properties:
        code: { type: string }
        message: { type: string }
        correlationId: { type: string, format: uuid }
        fieldErrors: { type: object, additionalProperties: { type: string } }
        recovery: { $ref: "#/components/schemas/Recovery" }
...
    JobDescription:
      type: object
      required: [id, sourceType, status, correctedVersion, version]
      properties:
        id: { type: string, format: uuid }
        sourceType: { type: string, enum: [PASTED_TEXT, PDF, IMAGE] }
        status: { type: string, enum: [DRAFT, EXTRACTING, READY_FOR_REVIEW, CONFIRMED, ANALYZED, FAILED] }
        correctedText: { type: [string, "null"] }
        correctedVersion: { type: integer }
        version: { type: integer }
...
    PageInfo:
      type: object
      required: [page, pageSize, total]
      properties: { page: { type: integer }, pageSize: { type: integer }, total: { type: integer } }
  responses:
    Success: { description: Successful resource response, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
    Page: { description: Paginated response, content: { application/json: { schema: { type: object, required: [items, pageInfo], properties: { items: { type: array, items: { type: object, additionalProperties: true } }, pageInfo: { $ref: "#/components/schemas/PageInfo" } } } } } }
    Error: { description: Safe actionable error, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
    JobDescription: { description: Private job description projection, content: { application/json: { schema: { $ref: "#/components/schemas/JobDescription" } } } }
...